create useBookData custom hooks
